make control bar independent block, with its on scss file and js file, that has code for creating this block. Change movieVars.js, and moviePageElem
