+++ controlBar.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./src/assets/styles/normalize.css" />
+    <link rel="stylesheet" href="./src/main.css" />
+    <title>Document</title>
+  </head>
+  <body>
+    <div class="example-container" id="pasteTo">
+      <!-- <div class="control-bar">
+        <button class="control-bar__btn" id="likeMovieBtn">
+          <img
+            src="./public/icons/likeIcon.svg"
+            alt="Like icon"
+            class="control-bar__icon"
+          />
+        </button>
+        <button class="control-bar__btn" id="addMovieBtn">
+          <img
+            src="./public/icons/plusIcon.svg"
+            alt="plus icon"
+            class="control-bar__icon"
+          />
+        </button>
+      </div> -->
+    </div>
+    <script type="module" src="./src/js/moviePage/createControlBar.js"></script>
+  </body>
+</html>
