+++ src/assets/styles/blocks/_controlBar.scss
@@ -1,0 +1,19 @@
+.control-bar {
+  display: flex;
+  gap: 10px;
+
+  &__btn {
+    padding: 14px;
+    width: 56px;
+    height: 56px;
+    border-radius: 8px;
+    border: 1px solid #262626;
+    background: #0f0f0f;
+  }
+
+  &__icon {
+    width: 28px;
+    height: 28px;
+    object-fit: contain;
+  }
+}
