+++ src/main.scss
@@ -11,3 +11,4 @@
 @import "./assets/styles/blocks/moviePage";
 @import "./assets/styles/blocks/_header.scss";
 
+@import "./assets/styles/blocks/controlBar";
